requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.11.7
python-dotenv==1.0.1
openai==1.82.0
pyyaml==6.0.2
